DB-4552: Restricted access if nonce is invlide.

+++ src/class-decoupled-preview-settings.php
@@ -109,11 +109,19 @@ if ( ! class_exists( 'Decoupled_Preview_Settings' ) ) {
 			if ( ! current_user_can( 'manage_options' ) ) {
 				wp_die( esc_html( 'You do not have sufficient permissions to access this page.' ) );
 			}
-			$edit_id = $this->get_nonce_verify( 'edit' );
-			if ( isset( $edit_id ) ) {
-				$action = 'options.php?edit=' . $edit_id . '&nonce=' . wp_create_nonce( 'edit_' . $edit_id );
-			} else {
+			$edit_id = $this->get_action_id( 'edit' );
+			if ( isset( $edit_id ) && isset( $_REQUEST['edit'] ) ) {
+				$action = add_query_arg(
+					[
+						'edit'  => $edit_id,
+						'nonce' => wp_create_nonce( 'edit' . $edit_id ),
+					],
+					'options.php'
+				);
+			} elseif ( ! isset( $edit_id ) && ! isset( $_REQUEST['edit'] ) ) {
 				$action = 'options.php';
+			} else {
+				wp_die( 'Unable perform action: invalid nonce' );
 			}
 
 			?>
@@ -125,9 +133,15 @@ if ( ! class_exists( 'Decoupled_Preview_Settings' ) ) {
 						<input name="Submit" type="submit" class="button-primary" value="<?php esc_attr_e( 'Save Changes' ); ?>" />
 						<?php
 						if ( isset( $edit_id ) ) {
-							$href = 'options-general.php?page=delete_preview_site&delete=' . $edit_id . '&nonce=' . wp_create_nonce( 'delete_' . $edit_id );
+							$url = add_query_arg(
+								[
+									'delete' => $edit_id,
+									'nonce'  => wp_create_nonce( 'delete' . $edit_id ),
+								],
+								'options-general.php?page=delete_preview_site'
+							);
 							?>
-							<a id="delete-preview" class="button-secondary button-large" href="<?php echo esc_html( $href ); ?>"><?php esc_attr_e( 'Delete' ); ?></a>
+							<a id="delete-preview" class="button-secondary button-large" href="<?php echo esc_html( $url ); ?>"><?php esc_attr_e( 'Delete' ); ?></a>
 							<?php
 						}
 						?>
@@ -146,7 +160,12 @@ if ( ! class_exists( 'Decoupled_Preview_Settings' ) ) {
 			if ( ! current_user_can( 'manage_options' ) ) {
 				wp_die( esc_html( 'You do not have sufficient permissions to access this page.' ) );
 			}
-			$delete_id = $this->get_nonce_verify( 'delete' );
+			$delete_id = $this->get_action_id( 'delete' );
+
+			if ( ! $delete_id && isset( $_REQUEST['delete'] ) ) {
+				wp_die( 'Unable perform action: invalid nonce' );
+			}
+
 			if ( $delete_id ) {
 				$this->delete_preview_site( $delete_id );
 				echo '<script type="text/javascript">window.location = "options-general.php?page=preview_sites"</script>';
@@ -191,13 +210,20 @@ if ( ! class_exists( 'Decoupled_Preview_Settings' ) ) {
 							$listing_data['label']        = $option['label'];
 							$listing_data['url']          = $option['url'];
 							$listing_data['preview_type'] = $option['preview_type'];
-							$nonce                        = wp_create_nonce( 'edit_' . $id );
+							$nonce                        = wp_create_nonce( 'edit' . $id );
+							$url                          = add_query_arg(
+								[
+									'edit'  => $id,
+									'nonce' => $nonce,
+								],
+								'/wp/wp-admin/options-general.php?page=add_preview_site'
+							);
 							if ( isset( $option['content_type'] ) ) {
 								$listing_data['content_type'] = ucwords( implode( ', ', $option['content_type'] ) );
 							} else {
 								$listing_data['content_type'] = 'Post, Page';
 							}
-							$listing_data['edit'] = "<a href='/wp/wp-admin/options-general.php?page=add_preview_site&edit={$id}&nonce={$nonce}'>Edit</a>";
+							$listing_data['edit'] = "<a href='{$url}'>Edit</a>";
 							?>
 
 							<tr>
@@ -246,7 +272,7 @@ if ( ! class_exists( 'Decoupled_Preview_Settings' ) ) {
 		 * @return array|array[]|false|mixed
 		 */
 		public function sanitize_callback_preview( array $input ) {
-			if ( $this->get_nonce_verify( 'delete' ) ) {
+			if ( $this->get_action_id( 'delete' ) ) {
 				return [ 'preview' => $input ];
 			} else {
 				$options = get_option( 'preview_sites' );
@@ -257,7 +283,11 @@ if ( ! class_exists( 'Decoupled_Preview_Settings' ) ) {
 					}
 				}
 
-				$edit_id  = $this->get_nonce_verify( 'edit' );
+				$edit_id = $this->get_action_id( 'edit' );
+				if ( ! $edit_id && isset( $_REQUEST['edit'] ) ) {
+					wp_die( 'Unable perform action: invalid nonce' );
+				}
+
 				$last_key = array_key_last( $options['preview'] );
 				if ( 1 === $last_key && null === $options['preview'][1]['label'] ) {
 					return [ 'preview' => [ 1 => $input ] ];
@@ -286,15 +316,15 @@ if ( ! class_exists( 'Decoupled_Preview_Settings' ) ) {
 		}
 
 		/**
-		 * Get the nonce verify and return the get value.
+		 * Get the nonce verify and return the action id.
 		 *
 		 * @param string $action Action name.
 		 *
 		 * @return mixed|null
 		 */
-		public function get_nonce_verify( string $action ) {
-			if ( isset( $_GET[ $action ] ) && isset( $_GET['nonce'] ) && wp_verify_nonce( wp_unslash( $_GET['nonce'] ), $action . '_' . $_GET[ $action ] ) ) {
-				return $_GET[ $action ];
+		public function get_action_id( string $action ) {
+			if ( isset( $_REQUEST[ $action ] ) && isset( $_REQUEST['nonce'] ) && wp_verify_nonce( wp_unslash( $_REQUEST['nonce'] ), $action . $_REQUEST[ $action ] ) ) {
+				return $_REQUEST[ $action ];
 			}
 			return null;
 		}
@@ -305,7 +335,7 @@ if ( ! class_exists( 'Decoupled_Preview_Settings' ) ) {
 		 * @return void
 		 */
 		public function setting_label_fn() {
-			$edit_id = $this->get_nonce_verify( 'edit' );
+			$edit_id = $this->get_action_id( 'edit' );
 			$site    = $this->get_preview_site( $edit_id );
 			$value   = isset( $edit_id ) ? $site['label'] : '';
 			echo wp_kses(
@@ -330,7 +360,7 @@ if ( ! class_exists( 'Decoupled_Preview_Settings' ) ) {
 		 * @return void
 		 */
 		public function setting_url_fn() {
-			$edit_id = $this->get_nonce_verify( 'edit' );
+			$edit_id = $this->get_action_id( 'edit' );
 			$site    = $this->get_preview_site( $edit_id );
 			$value   = isset( $edit_id ) ? $site['url'] : '';
 			echo wp_kses(
@@ -355,7 +385,7 @@ if ( ! class_exists( 'Decoupled_Preview_Settings' ) ) {
 		 * @return void
 		 */
 		public function setting_secret_fn() {
-			$edit_id = $this->get_nonce_verify( 'edit' );
+			$edit_id = $this->get_action_id( 'edit' );
 			$html    = isset( $edit_id ) ? "<input id='plugin_text_secret' name='preview_sites[secret_string]' size='40' type='password' /><br>Shared secret for the preview site. When editing, if kept empty the old value will be saved, otherwise it will be overwritten." : "<input id='plugin_text_secret' name='preview_sites[secret_string]' size='40' type='password' required /><br>[Required] Shared secret for the preview site.";
 			echo wp_kses(
 				$html,
@@ -379,7 +409,7 @@ if ( ! class_exists( 'Decoupled_Preview_Settings' ) ) {
 		 * @return void
 		 */
 		public function setting_preview_type_fn() {
-			$edit_id = $this->get_nonce_verify( 'edit' );
+			$edit_id = $this->get_action_id( 'edit' );
 			$site    = $this->get_preview_site( $edit_id );
 			$items   = array( 'Next.js' );
 			echo wp_kses(
@@ -423,7 +453,7 @@ if ( ! class_exists( 'Decoupled_Preview_Settings' ) ) {
 		 * @return void
 		 */
 		public function setting_content_type_fn() {
-			$edit_id = $this->get_nonce_verify( 'edit' );
+			$edit_id = $this->get_action_id( 'edit' );
 			$items   = array( 'Post', 'Page' );
 			$site    = $this->get_preview_site( $edit_id );
 			foreach ( $items as $item ) {
